+++ src/Armature.Core/src/BuildChainPatterns/IBuildChainPattern.cs
@@ -24,10 +24,10 @@ public interface IBuildChainPattern : IEquatable<IBuildChainPattern>, ILogString
   ///   The chain of units representing a build session, the last one is the unit to be built,
   ///   the previous are the context of the build session. Each next unit is the dependency of the previous one.
   /// </param>
+  /// <param name="actionBag"></param>
   /// <param name="inputWeight">
   ///   The weight of matching which used by children matchers to calculate a final weight of matching
   /// </param>
-  /// <param name="actionBag"></param>
   /// <remarks>
   /// If there is type A which depends on class B, during building A, B should be built and the build chain will be [A, B] in this case.
   /// </remarks>
@@ -35,5 +35,5 @@ public interface IBuildChainPattern : IEquatable<IBuildChainPattern>, ILogString
   /// Returns all matched build actions for the <paramref name="buildChain" />. All actions are grouped by a building stage
   /// and coupled with a "weight of matching". See <see cref="WeightedBuildActionBag" /> for details.
   /// </returns>
-  bool GatherBuildActions(BuildChain buildChain, out WeightedBuildActionBag? actionBag, int inputWeight = 0);
+  bool GatherBuildActions(BuildChain buildChain, out WeightedBuildActionBag? actionBag, long inputWeight = 0);
 }
