+++ tests/Tests/src/UnitTests/BuildChainPatterns/IfFirstUnitTest.cs
@@ -28,10 +28,10 @@ public class IfFirstUnitTest
     target.GatherBuildActions(chain, out var actionBag, 0);
 
     // --assert
-    A.CallTo(() => child1.GatherBuildActions(An<BuildChain>.That.IsEqualTo(expected.ToBuildChain(), Comparer.OfArrayTail<UnitId>()), out actionBag, An<int>._))
+    A.CallTo(() => child1.GatherBuildActions(An<BuildChain>.That.IsEqualTo(expected.ToBuildChain(), Comparer.OfArrayTail<UnitId>()), out actionBag, An<long>._))
      .MustHaveHappenedOnceAndOnly();
 
-    A.CallTo(() => child2.GatherBuildActions(An<BuildChain>.That.IsEqualTo(expected.ToBuildChain(), Comparer.OfArrayTail<UnitId>()), out actionBag, An<int>._))
+    A.CallTo(() => child2.GatherBuildActions(An<BuildChain>.That.IsEqualTo(expected.ToBuildChain(), Comparer.OfArrayTail<UnitId>()), out actionBag, An<long>._))
      .MustHaveHappenedOnceAndOnly();
   }
